Update README.md
This commit introduces the initial implementation of the Auto-Bangka-Transfer.py script. The software provides functionality for converting text into a binary representation using UTF-8 encoding, with the binary digits being represented as the custom characters 'Bangka' (邦咔). It also includes a decoding feature, which allows users to convert 'Bangka' binary sequences back into their original text format. This tool was inspired by the character Alice Tendo from Blue Archive and was developed as a fun and creative way to encode and decode messages. Additionally, the script automatically copies the encoded result to the clipboard for user convenience.